impl(internal/parser): move genclient/common.go
Now that protobuf and openapi are in the same package, move
genclient.SplitApiName and genclient.WellKnownMixin to package parser
and unexport them, since that is the open place these function are used.
