+++ generator/internal/parser/protobuf.go
@@ -224,9 +224,9 @@ func makeAPIForProtobuf(serviceConfig *serviceconfig.Service, req *pluginpb.Code
 		enabledMixinMethods, mixinFileDesc = loadMixins(serviceConfig)
 		packageName := ""
 		for _, api := range serviceConfig.Apis {
-			packageName, _ = genclient.SplitApiName(api.Name)
+			packageName, _ = splitApiName(api.Name)
 			// Keep searching after well-known mixin services.
-			if !genclient.WellKnownMixin(api.Name) {
+			if !wellKnownMixin(api.Name) {
 				break
 			}
 		}
